requirements.txt
click
requests
xmltodict
dogpile.cache
